% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Circ.ratioplot.R
\name{Circ.ratioplot}
\alias{Circ.ratioplot}
\title{Circ.ratioplot}
\usage{
Circ.ratioplot(Circ, Linear, CircCoordinates, plotrow = "1", size = 18,
  ncol = 2, groupindicator1 = NULL, groupindicator2 = NULL,
  x = "Conditions", y = "circRNA/(circRNA+Linear)",
  lab_legend = "groupindicator1")
}
\arguments{
\item{Circ}{CircRNACount file. A file of circRNA read count table. First three columns are circRNA coordinates, and followed by columns for circRNA read counts, each sample per column.}

\item{Linear}{LinearCount file. A file of circRNA host gene expression count table. Same configuration as CircRNACount file.}

\item{CircCoordinates}{BED format circRNA coordinates file.}

\item{plotrow}{The rownumber or rowname in the CircRNACount table corresponding the specific gene which you want to plot.}

\item{size}{the text size}

\item{ncol}{if groupindicator2 is provided, specify the panel layout.}

\item{groupindicator1}{A vector of group indicators. For example, ages.}

\item{groupindicator2}{An other vector of group indicators. For example, tissues. This indicator will be used to segement plots out.}

\item{x}{x axis lable}

\item{y}{y axis lable}

\item{lab_legend}{legend title}
}
\description{
Plot the ratio of circRNAs to the total transcripts (circRNA plus liner RNA).
}
\examples{
data(Coordinates)
data(Circ)
data(Linear)
Circ.ratioplot( Circ, Linear, Coordinates, plotrow=10, 
                groupindicator1=c(rep('1days',6),rep('4days',6),rep('20days',6)),
                groupindicator2=rep(c(rep('Female',4),rep('Male',2)),3), 
                lab_legend='Ages' )
}

