% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{Coordinates}
\alias{Coordinates}
\title{Coordinates}
\format{A data frame with 153 rows and 6 variables.
\describe{
 \item{Chr}{Chromosomes of CircRNAs}
 \item{Start}{Start positions of CircRNAs}
 \item{End}{End positions of CircRNAs}
 \item{gene}{gene name}
 \item{junctiontype}{junction type. 1 and 2 are GT-AG and reverse complementary. 0 is non canonical junction}
 \item{strand}{strand of circRNA}
}}
\description{
Coordinates
}

