% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostGeneCount.R
\name{HostGeneCount}
\alias{HostGeneCount}
\title{HostGeneCount}
\usage{
HostGeneCount(CircCoordinates, bamfiles, ncore = 2, stranded = "unstranded",
  splicedOnly = TRUE)
}
\arguments{
\item{CircCoordinates}{CircCoordinates output of DCC, or any other BED6 format files, input as data.frame}

\item{bamfiles}{A iterable list of bamfiles with full path.}

\item{ncore}{Number of core to use.}

\item{stranded}{Whether the library is prepared strand specific, default "unstranded". Accepted values are "unstranded","fs","ss", meaning "unstranded library", "first strand library" and "second strand library" respectively.}

\item{splicedOnly}{Only count reads that are spliced at the circRNA junction, default TRUE.}
}
\description{
Count circRNA host gene expression from given BED6 format CircCoordinates.
}

