% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostGeneCount.R
\name{read_one_scanBam}
\alias{read_one_scanBam}
\title{Read one element of the result of sanBam function from Rsamtools. This function Called by read_scanBam.}
\usage{
read_one_scanBam(scanBam_one_result, strand = "+", splicedOnly = TRUE)
}
\arguments{
\item{scanBam_one_result}{one element of the sanBam result, corresponding to one genomic region.}

\item{strand}{of the region, "+", "-" or "*".}

\item{splicedOnly}{Only count reads that are spliced at the circRNA junction. Default to be TRUE.}
}
\description{
Read one element of the result of sanBam function from Rsamtools. This function Called by read_scanBam.
}

