% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostGeneCount.R
\name{read_scanBam}
\alias{read_scanBam}
\title{Read scanBam ouput list object.}
\usage{
read_scanBam(scanBam_result, strand = strands, splicedOnly = TRUE)
}
\arguments{
\item{scanBam_result}{the result list object generated by Rsamtools::scanBam function}

\item{strand}{a vector of strand information, need to be the same length as the "which" input option of sanBam (the number of circRNA subjected to host gene counting).
CircTest needs this information to determine which strand of the reads should be counted.}

\item{splicedOnly}{Only count reads that are spliced at the circRNA junction. Default to be TRUE.}
}
\description{
Read scanBam ouput list object.
}

